+++ custom_model_runner/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.5.4] - 2021-04-13
+##### Added
+- parameter support for training models and transformers
 
 #### [1.5.3] - 2021-04-03
 ##### Fixed